feat : mypage dto 수정

See merge request s09-final/S09P31B303!233

+++ greenstep/src/main/java/com/mm/greenstep/domain/mypage/dto/response/MyPageAllPloggingResDto.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 public class MyPageAllPloggingResDto {
     private Double travelRange;
-    private Double travelTime;
+    private Long travelTime;
     private Integer trashAmount;
     private Integer completedAchieveCount;
 }
